requirements.txt
pandas
requests
sqlalchemy
pymysql
python-dotenv
prefect
fuzzywuzzy[speedup]
pyarrow
# optional:
# streamlit